@@ -7,11 +7,14 @@ dependencies:
   flutter:
     sdk: flutter
 
+environment:
+  sdk: ">=2.12.0 <3.0.0"
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  e2e: ^0.5.0
-  pigeon: ^0.1.9
+  e2e: ^0.7.0+1
+  pigeon: ^1.0.1
   flutter_driver:
     sdk: flutter
 
@@ -24,6 +27,4 @@ flutter:
       ios:
         pluginClass: ZendeskPlugin
 
-environment:
-  sdk: ">=2.8.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.6 <2.0.0"
+
